requirements.txt
h5py==3.1.0
pytest
wandb
boto3
spotty
torch-geometric
python-dotenv
click